requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.3

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Embeddings & Vector Stores
sentence-transformers==2.2.2
chromadb==0.4.24
faiss-cpu==1.7.4

# LLMs & Generation
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.13

# Web & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0

# Database & Logging
sqlite3-python==1.0.0
logging==0.5.1.2

# Utilities
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
requests==2.31.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3
